requirements.txt
requests~=2.32.4
pymongo~=4.13.2
qdrant-client~=1.14.3
cohere~=5.15.0
pandas~=2.3.0
fastapi~=0.115.13
python-dotenv~=1.1.0
uvicorn~=0.29.0
python-multipart~=0.0.6
